LU-12898 utils: %llu mismatch with type __u64 on ppcle64
Fix build errors like this one on ppcle64:

BUILDSTDERR: libmount_utils_zfs.c: In function 'zfs_mkfs_opts':
BUILDSTDERR: libmount_utils_zfs.c:573:5: error: format '%llu' expects
argument of type 'long long unsigned int', but argument 4 has type
'__u64' [-Werror=format=]
BUILDSTDERR:      mop->mo_device_kb * 1024);

__u64 was treated as an unsigned long long which breaks the build on
ppc64le, where they are not the same size.

In printf cases, cast to uintmax_t, and change the printf format to
"%ju", so the format is compatible with the type and it is guaranteed
not to lose any data.

In the case of sscanf(), replace the call with strtoull() to eliminate
the issue.

+++ lustre/utils/libmount_utils_zfs.c
@@ -31,6 +31,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <libzfs.h>
+#include <inttypes.h>
 #include <sys/systeminfo.h>
 
 /* Persistent mount data is stored in these user attributes */
@@ -569,8 +570,8 @@ static char *zfs_mkfs_opts(struct mkfs_opts *mop, char *str, int len)
 	if (strlen(mop->mo_mkfsopts) != 0)
 		snprintf(str, len, " -o %s", mop->mo_mkfsopts);
 	if (mop->mo_device_kb)
-		snprintf(str, len, " -o quota=%llu",
-			 mop->mo_device_kb * 1024);
+		snprintf(str, len, " -o quota=%ju",
+			 (uintmax_t) mop->mo_device_kb * 1024);
 
 	return str;
 }
